chore: address swiftlint --strict violations
CI runs `swiftlint --strict` and our PR introduced four classes of
violations:

* `PosixCompat+Windows.swift` — POSIX-shaped names (`mode_t`,
  `S_IFMT`, `timegm`, …) intentionally match the libc spelling so
  call sites read identically across platforms; bracket the file
  with `swiftlint:disable identifier_name type_name` /
  `swiftlint:enable` and rename two short closure params (`tm`,
  `fd`) that aren't ABI-visible. Also fix the double-space colon
  spacing on the `S_IF*` constants.

* `FileManager+ZIP.swift` was over the 400-line file_length cap
  after the Windows fallbacks were inlined per-method. Lift the
  Windows variants of `permissionsForItem(at:)`,
  `fileModificationDateTimeForItem(at:)`, `fileSizeForItem(at:)`,
  and `typeForItem(at:)` into a new `FileManager+ZIPWindows.swift`
  and gate the POSIX implementations with `#if !os(Windows)`.

* `Entry.swift` was also over the cap; move the 256-entry
  `cp437Lookup` table out into a new `Entry+CodePage437.swift`
  (used only on Linux/Windows/Android, where there's no
  `CFStringEncoding`).

* `Package.swift` had a trailing comma in the new `linkerSettings`
  array.

Pure refactor — no behaviour change on any platform.
